pyproject.toml
[tool.poetry]
name = "hugo2zola"
version = "0.1.0"
description = "Migrate a Hugo site to Zola"
authors = ["Chris Proctor <github.com@accounts.chrisproctor.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-frontmatter = "^1.0.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
